app/src/main/java/com/antonioleiva/bandhookkotlin/data/mapper/ImageMapper.kt
package com.antonioleiva.bandhookkotlin.data.mapper

import com.antonioleiva.bandhookkotlin.data.lastfm.model.LastFmImage
import com.antonioleiva.bandhookkotlin.data.lastfm.model.LastFmImageType

/**
 * @author tpom6oh@gmail.com
 *
 * 03/07/16.
 */

class ImageMapper {
    fun getMainImageUrl(images: List<LastFmImage>?): String? {
        if (images == null || images.isEmpty()) {
            return null
        }

        val image = images.firstOrNull { it.size == LastFmImageType.MEGA.type }
        return image?.url ?: images.last().url
    }
}